Criação de service para atendente, alteração de get do tipo integer para long

+++ src/test/java/com/trabalhoFinal/trabalhoFinal/models/Atendente.java
@@ -42,7 +42,7 @@ public class Atendente implements Serializable {
     public Atendente() {
     }
 
-    public Integer getCodAtendente() {
+    public Long getCodAtendente() {
         return codAtendente;
     }
